add render fn
