+++ package.json
@@ -23,7 +23,7 @@
     "dependencies": {
         "@figma-plugin/helpers": "^0.15.2",
         "@octokit/rest": "^18.9.0",
-        "@radix-ui/react-checkbox": "^0.1.1",
+        "@radix-ui/react-checkbox": "^0.1.4",
         "@radix-ui/react-context-menu": "^0.0.23",
         "@radix-ui/react-dropdown-menu": "^0.0.19",
         "@radix-ui/react-icons": "^1.0.3",
