+++ src/app/components/Checkbox.tsx
@@ -53,8 +53,6 @@ function Checkbox({
       <StyledIndicator>
         <CheckIcon />
       </StyledIndicator>
-      {checked ? 'CHECK' : 'UNCHECK'}
-      {defaultChecked ? 'defCHECK' : 'defUNCHECK'}
     </StyledCheckbox>
   );
 }
