+++ src/app/components/InspectorMultiView.tsx
@@ -11,6 +11,69 @@ import Heading from './Heading';
 import Label from './Label';
 import Blankslate from './Blankslate';
 
+function renderResolvedtoken(token) {
+  if (!token) return null;
+  switch (token?.type) {
+    case 'color': {
+      return (
+        <Box
+          css={{
+            background: token.value,
+            width: '24px',
+            height: '24px',
+            borderRadius: '100%',
+          }}
+        />
+      );
+    }
+    case 'typography': {
+      return (
+        <Box
+          css={{
+            background: '$bgSubtle',
+            fontSize: '$small',
+            padding: '$2 $3',
+            borderRadius: '$default',
+            width: '40px',
+          }}
+        >
+          aA
+        </Box>
+      );
+    }
+    case 'boxShadow': {
+      return (
+        <Box
+          css={{
+            background: '$bgSubtle',
+            fontSize: '$small',
+            padding: '$2 $3',
+            borderRadius: '$default',
+            width: '40px',
+          }}
+        >
+          shd
+        </Box>
+      );
+    }
+    default: {
+      return (
+        <Box
+          css={{
+            background: '$bgSubtle',
+            fontSize: '$small',
+            padding: '$2 $3',
+            borderRadius: '$default',
+            width: '40px',
+          }}
+        >
+          {token.value}
+        </Box>
+      );
+    }
+  }
+}
+
 export default function InspectorMultiView() {
   const uiState = useSelector((state: RootState) => state.uiState);
   const { findToken, removeTokensByValue } = useTokens();
